requirements.txt
cerebras_cloud_sdk
discord.py
aiohttp
python-dotenv
requests
websocket-client
pollinations.ai
colorama
logtail-python
PyNaCl
gTTS
vosk
mysql-connector-python